시험감독 변수 타입 수정

+++ BaekJoon_13458/src/Main.java
@@ -6,7 +6,7 @@ public class Main {
 	static int n;
 	static int[] array;
 	static int b,c;
-	static int ans;
+	static long ans;
 	
 	
 	public static void solving() {
